# apps/dashboard/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Marlin API",
    "description": "Stablecoin billing API for Solana. Invoices, subscriptions, customers, and webhooks.",
    "version": "1.0.0",
    "contact": { "name": "Marlin", "url": "https://marlin.fi" },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
  "servers": [
    { "url": "https://api.marlin.dev", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "security": [{ "BearerAuth": [] }],
...
  "paths": {
...
    "/api/subscriptions": {
      "get": {
        "tags": ["Subscriptions"],
        "summary": "List subscriptions",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["Active", "PastDue", "Cancelled", "Completed"] } },
          { "name": "planId", "in": "query", "schema": { "type": "string" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubscriptionList" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Subscriptions"],
        "summary": "Create subscription",
        "description": "Creates a subscription for a customer to a plan (merchant-initiated).",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["planId", "customerWallet"], "properties": { "planId": { "type": "string" }, "customerWallet": { "type": "string" } } } } }
        },
        "responses": {
          "201": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subscription" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/webhooks": {
      "post": {
        "tags": ["Settings"],
        "summary": "Register webhook endpoint",
        "description": "Registers or updates the webhook URL for receiving event notifications.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["url"], "properties": { "url": { "type": "string", "format": "uri" } } } } }
        },
        "responses": {
          "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookConfig" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "ApiKeyAuth": { "type": "apiKey", "in": "header", "name": "Authorization", "description": "Bearer mk_live_..." }
    },
    "responses": {
      "Unauthorized": { "description": "Authentication required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "ValidationError": { "description": "Invalid input", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object" }
            }
          }
        }
      },
...
      "Subscription": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "onchainId": { "type": "string" },
          "planId": { "type": "string" },
          "customerId": { "type": "string" },
          "status": { "type": "string", "enum": ["Active", "PastDue", "Cancelled", "Completed"] },
          "currentPeriodStart": { "type": "string", "format": "date-time" },
          "currentPeriodEnd": { "type": "string", "format": "date-time" },
          "cancelledAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "SubscriptionList": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Subscription" } },
          "has_more": { "type": "boolean" },
          "cursor": { "type": "string", "nullable": true }
        }
      },
...
      "WebhookConfig": {
        "type": "object",
        "properties": {
          "url": { "type": "string", "nullable": true },
          "secret": { "type": "string", "description": "Only returned on creation/rotation" }
        }
      },
...
    }
  }
}